pass necessary functions to executors
